prevent trying to copy / if for some reason something goes wrong

+++ runescape/rsu/framework/modules/rsu/files/copy.pm
@@ -17,38 +17,43 @@ sub print_cpr
 	# Get the passed data
 	my ($from,$to,$replace) = @_;
 	
-	# If replacing content was requested
-	if (defined $replace && $replace =~ /^(1|true)$/i)
+	# If the $from is not /
+	if ($from !~ /^\/$/)
 	{
-		# Tell user what we are doing
-		print "Replacing content in:\n\"$to/\"\nWith content from:\n\"$from/\"\n";
-		
-		if ($OS !~ /darwin/)
+		# If replacing content was requested
+		if (defined $replace && $replace =~ /^(1|true)$/i)
 		{
-			# Enable Remove Target Directory Before Copy
-			local $File::Copy::Recursive::RMTrgDir = 2;
+			# Tell user what we are doing
+			print "Replacing content in:\n\"$to/\"\nWith content from:\n\"$from/\"\n";
+			
+			if ($OS !~ /darwin/)
+			{
+				# Enable Remove Target Directory Before Copy
+				local $File::Copy::Recursive::RMTrgDir = 2;
+				
+				# Copy $from to $to
+				dircopy($from, $to) or warn $!;
+			}
+			# Else
+			else
+			{
+				# Make the path to where we copy the file
+				make_path($to);
+				
+				# Copy using rsync
+				system "rsync -r --delete \"$from/\"* \"$to\"";
+			}
+		}
+		else
+		{
+			# Tell user what we are doing
+			print "cp: \"$from/\" -> \"$to/\"\n";
 			
 			# Copy $from to $to
 			dircopy($from, $to) or warn $!;
 		}
-		# Else
-		else
-		{
-			# Make the path to where we copy the file
-			make_path($to);
-			
-			# Copy using rsync
-			system "rsync -r --delete \"$from/\"* \"$to\"";
-		}
 	}
-	else
-	{
-		# Tell user what we are doing
-		print "cp: \"$from/\" -> \"$to/\"\n";
-		
-		# Copy $from to $to
-		dircopy($from, $to) or warn $!;
-	}
+	
 }
 
 #
@@ -60,38 +65,42 @@ sub print_mvr
 	# Get the passed data
 	my ($from,$to,$replace) = @_;
 	
-	# If replacing content was requested
-	if (defined $replace && $replace =~ /^(1|true)$/i)
+	# If the $from is not /
+	if ($from !~ /^\/$/)
 	{
-		# Tell user what we are doing
-		print "Replacing content in:\n\"$to/\"\nWith content from:\n\"$from/\"\n";
-		
-		# If we are not on mac osx
-		if ($OS !~ /darwin/)
+		# If replacing content was requested
+		if (defined $replace && $replace =~ /^(1|true)$/i)
 		{
-			# Enable Remove Target Directory Before Copy
-			local $File::Copy::Recursive::RMTrgDir = 2;
-		
+			# Tell user what we are doing
+			print "Replacing content in:\n\"$to/\"\nWith content from:\n\"$from/\"\n";
+			
+			# If we are not on mac osx
+			if ($OS !~ /darwin/)
+			{
+				# Enable Remove Target Directory Before Copy
+				local $File::Copy::Recursive::RMTrgDir = 2;
+			
+				# Copy $from to $to
+				dirmove($from, $to) or warn $!;
+			}
+			# Else
+			else
+			{
+				# Make the path to where we copy the file
+				make_path($to);
+				
+				# Copy using rsync
+				system "rsync -r --delete \"$from/\"* \"$to\"";
+			}
+		}
+		else
+		{
+			# Tell user what we are doing
+			print "mv: \"$from/\" -> \"$to/\"\n";
+			
 			# Copy $from to $to
 			dirmove($from, $to) or warn $!;
 		}
-		# Else
-		else
-		{
-			# Make the path to where we copy the file
-			make_path($to);
-			
-			# Copy using rsync
-			system "rsync -r --delete \"$from/\"* \"$to\"";
-		}
-	}
-	else
-	{
-		# Tell user what we are doing
-		print "mv: \"$from/\" -> \"$to/\"\n";
-		
-		# Copy $from to $to
-		dirmove($from, $to) or warn $!;
 	}
 }
